refactor(channel-store-admin): extract transport health helper module (#1293)
